wf(android-sdk-3-2-bucketing-sticky): review R1 — JS SDK parity tightening
Fixes from code review round 1:

1. BucketingManager.selectBucket accumulator: Int -> Double. JS SDK's
   cumulative-percentage math uses double-precision throughout (pct*100 +
   redistribute is a Number). Fractional percentages like 33.333 were
   silently truncated by the Int accumulator, causing off-by-one bucket
   selection around boundary values. Story 3.5's 50+ parity vectors
   would have caught this — closing it early.

2. Variation.bucketingAllocation field semantics: was 'configured traffic
   percentage', now 'hash-pipeline value 0..maxTraffic'. Matches JS SDK
   BucketedVariation.bucketingAllocation. The sticky path yields null
   (no fresh hash was computed) — matches JS SDK's sticky branch that
   leaves the field undefined. Variation.kt KDoc corrected to match.

3. Two tests updated to compare identity fields (id/key/experienceId)
   instead of whole-object equality since the fresh vs sticky Variation
   objects now differ on bucketingAllocation.

+++ packages/core/src/main/kotlin/com/convert/sdk/core/bucketing/BucketingManager.kt
@@ -49,10 +49,11 @@ import com.goncalossilva.murmurhash.MurmurHash3
  * ### Bucket selection (AC-3)
  *
  * `selectBucket` walks the `buckets` map in insertion order, accumulating
- * `pct * 100 + redistribute` into `prev`. The first bucket whose running
- * total strictly exceeds `value` wins. A value at or past the sum of all
- * boundaries returns `null`, signalling "not bucketed" (common when
- * `max_traffic < 10000`).
+ * `pct * 100 + redistribute` into `prev` (typed as [Double] to preserve
+ * fractional boundaries for non-integer percentages — JS SDK parity). The
+ * first bucket whose running total strictly exceeds `value` wins. A value
+ * at or past the sum of all boundaries returns `null`, signalling "not
+ * bucketed" (common when `max_traffic < 10000`).
  *
  * ### Why insertion order matters
  *
@@ -168,10 +169,16 @@ public class BucketingManager(
         value: Int,
         redistribute: Int = 0,
     ): String? {
-        var prev = 0
+        // prev is Double, not Int — the JS SDK's `prev += buckets[id] * 100 +
+        // redistribute` runs on JS Numbers (double-precision floats), so a
+        // fractional percentage like `33.333` yields fractional boundaries
+        // (3333.3, 6666.6, 9999.9) that an Int accumulator would silently
+        // truncate. Preserving Double keeps the comparison `value < prev`
+        // byte-identical to the JS SDK output on non-integer percentages.
+        var prev = 0.0
         var selected: String? = null
         for ((variationId, pct) in buckets) {
-            prev += (pct * PERCENTAGE_TO_BASIS_MULTIPLIER).toInt() + redistribute
+            prev += pct * PERCENTAGE_TO_BASIS_MULTIPLIER + redistribute
             if (value < prev) {
                 selected = variationId
                 break
@@ -235,7 +242,11 @@ public class BucketingManager(
          * Multiplier applied to each bucket's percentage (0..100) to lift
          * it into the `0..10000` basis-point space that [selectBucket]
          * accumulates against. Matches JS SDK `buckets[id] * 100`.
+         *
+         * Typed as [Double] (not [Int]) so the `pct * MULT + redistribute`
+         * expression retains the fractional component for non-integer
+         * percentages — see the Double-typed accumulator in [selectBucket].
          */
-        private const val PERCENTAGE_TO_BASIS_MULTIPLIER: Int = 100
+        private const val PERCENTAGE_TO_BASIS_MULTIPLIER: Double = 100.0
     }
 }
